+++ constants/botList.py
@@ -1,0 +1,3 @@
+known_bot_list = [
+    'streamelements'
+]