+++ constants/pop_culture_quotes.py
@@ -1,0 +1,3 @@
+quotes = {
+    'hello there': 'General Kenobi!'
+}