+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 21/02/2023
 
+# version
+
+- moved messages to contants folder
+- added twitch url to env file
+- add bot ignore list
+- added random welcome back message
+
+## 21/02/2023
+
 # version 1.0.1
 
 removed unwanted print commands